+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 * added in indicators in week/events if it is today/this week
 * added in current_week property in schedule class
 * added in events_today property in weeks class
+* added in function to event to load the officials for it
+* added in function to add broadcast info to event
 * 
 
 ## 0.3.3